requirements.txt
pandas==1.5.1
pyarrow==10.0.0
plotly==5.11.0
streamlit==1.14.0
google-cloud-storage==2.6.0
gcsfs==2022.11.0